+++ test/chain.unit.spec.js
@@ -1,6 +1,7 @@
 const assert = require('chai').assert,
     { Entry } = require('../src/entry'),
-    { Chain, composeChainCommit, composeChainReveal, composeChain, computeChainTxId } = require('../src/chain');
+    sign = require('tweetnacl/nacl-fast').sign,
+    { Chain, composeChainCommit, composeChainReveal, composeChain, composeChainLedger, computeChainTxId } = require('../src/chain');
 
 describe('Test Chain', function() {
 
@@ -61,6 +62,42 @@ describe('Test Chain', function() {
             '000162a772f640448ee02e500a8e539bcf5c02e53f8b88fc1f81f0a87d0f18af94ab9384992b5c9db26ac6b20aba137815efc39cf19cee53de0baccc54ec4e6acc6b02ffe4b936b56a6bbae773f5b51001161efdeb0ba1e8447f3c45206119b40539e4325cc5be0b5d54e4b02234a10b542573645f7ba55650f25eb931985cddcf451df77594b5b6a523cd0ebb71b13ec133eeb93c084958f1ed6adef51ec9ec6323c543f91303739c9e5194972c5105929b7787d327755ab1b5cf1b2884d4877b8fdcbca1cfb00b');
     });
 
+    it('should compose Chain commit with provided signature', function() {
+        const entry = Entry.builder()
+            .extId('my ext id 1784465577795', 'utf8')
+            .content('first')
+            .timestamp(1523227752000)
+            .build();
+
+        const chain = new Chain(entry);
+
+        const commit = composeChainCommit(chain, 'EC2UFobcsWom2NvyNDN67Q8eTdpCQvwYe327ZeGTLXbYaZ56e9QR', 'a523cd0ebb71b13ec133eeb93c084958f1ed6adef51ec9ec6323c543f91303739c9e5194972c5105929b7787d327755ab1b5cf1b2884d4877b8fdcbca1cfb00b');
+        assert.instanceOf(commit, Buffer);
+        assert.equal(commit.toString('hex'),
+            '000162a772f640448ee02e500a8e539bcf5c02e53f8b88fc1f81f0a87d0f18af94ab9384992b5c9db26ac6b20aba137815efc39cf19cee53de0baccc54ec4e6acc6b02ffe4b936b56a6bbae773f5b51001161efdeb0ba1e8447f3c45206119b40539e4325cc5be0b5d54e4b02234a10b542573645f7ba55650f25eb931985cddcf451df77594b5b6a523cd0ebb71b13ec133eeb93c084958f1ed6adef51ec9ec6323c543f91303739c9e5194972c5105929b7787d327755ab1b5cf1b2884d4877b8fdcbca1cfb00b');
+    });
+
+    it('should reject invalid signature manually provided for chain commit (timestamp not fixed)', function () {
+        const entry = Entry.builder()
+            .extId('my ext id 1784465577795', 'utf8')
+            .content('first')
+            .build();
+
+        const chain = new Chain(entry);
+
+        const buffer = composeChainLedger(chain);
+        const secretKey = sign.keyPair.fromSeed(Buffer.from('19c71085f906291070d07cb64cf160692c033d5e6ac05342f2219c03b4a5c8c6', 'hex')).secretKey;
+        const signature = sign.detached(buffer, secretKey);
+
+        try {
+            composeChainCommit(chain, 'EC2UFobcsWom2NvyNDN67Q8eTdpCQvwYe327ZeGTLXbYaZ56e9QR', signature);
+        } catch (e) {
+            assert.instanceOf(e, Error);
+            return;
+        }
+        throw new Error('Should have rejected invalid signature');
+    });
+
     it('should compose Chain reveal', function() {
         const entry = Entry.builder()
             .chainId('954d5a49fd70d9b8bcdb35d252267829957f7ef7fa6c74f88419bdc5e82209f4')
@@ -83,6 +120,22 @@ describe('Test Chain', function() {
         const chain = new Chain(entry);
 
         const composed = composeChain(chain, 'Es2d1a3uPx7o5uXHmsCnSEK2EKatPA56n8RUFmW9uRrpPRBuk5bZ');
+        assert.instanceOf(composed.commit, Buffer);
+        assert.instanceOf(composed.reveal, Buffer);
+        assert.equal(composed.commit.toString('hex'),
+            '000162a772f640448ee02e500a8e539bcf5c02e53f8b88fc1f81f0a87d0f18af94ab9384992b5c9db26ac6b20aba137815efc39cf19cee53de0baccc54ec4e6acc6b02ffe4b936b56a6bbae773f5b51001161efdeb0ba1e8447f3c45206119b40539e4325cc5be0b5d54e4b02234a10b542573645f7ba55650f25eb931985cddcf451df77594b5b6a523cd0ebb71b13ec133eeb93c084958f1ed6adef51ec9ec6323c543f91303739c9e5194972c5105929b7787d327755ab1b5cf1b2884d4877b8fdcbca1cfb00b');
+        assert.equal(composed.reveal.toString('hex'), '00fcfe632c6ab1a7c71448e256e0487c4cfc34ceac90e997de8c4fdf8485e9a0fd001900176d79206578742069642031373834343635353737373935');
+    });
+
+    it('should compose Chain commit and reveal', function() {
+        const entry = Entry.builder()
+            .extId('my ext id 1784465577795', 'utf8')
+            .content('first')
+            .timestamp(1523227752000)
+            .build();
+        const chain = new Chain(entry);
+
+        const composed = composeChain(chain, 'EC2UFobcsWom2NvyNDN67Q8eTdpCQvwYe327ZeGTLXbYaZ56e9QR', 'a523cd0ebb71b13ec133eeb93c084958f1ed6adef51ec9ec6323c543f91303739c9e5194972c5105929b7787d327755ab1b5cf1b2884d4877b8fdcbca1cfb00b');
         assert.instanceOf(composed.commit, Buffer);
         assert.instanceOf(composed.reveal, Buffer);
         assert.equal(composed.commit.toString('hex'),