+++ test/entry.unit.spec.js
@@ -1,9 +1,10 @@
 const assert = require('chai').assert,
-    { Entry, composeEntryCommit, composeEntryReveal, composeEntry, computeEntryTxId } = require('../src/entry');
+    sign = require('tweetnacl/nacl-fast').sign,
+    { Entry, composeEntryCommit, composeEntryReveal, composeEntry, computeEntryTxId, composeEntryLedger } = require('../src/entry');
 
-describe('Test Entry', function() {
+describe('Test Entry', function () {
 
-    it('should populate Entry attributes', function() {
+    it('should populate Entry attributes', function () {
         const entry = Entry.builder()
             .chainId('cfb5d93e747d20433e3b14603f90a5eb152d0399e7278f9671ecf9763f8780e8')
             .extId('extId', 'utf8')
@@ -26,7 +27,7 @@ describe('Test Entry', function() {
         assert.equal(entry.timestamp, 1523072354);
     });
 
-    it('should get sizes with ext ids', function() {
+    it('should get sizes with ext ids', function () {
         const entry = Entry.builder()
             .extId('extId', 'utf8')
             .extId('extId2', 'utf8')
@@ -39,7 +40,7 @@ describe('Test Entry', function() {
 
     });
 
-    it('should copy Entry', function() {
+    it('should copy Entry', function () {
         const entry = Entry.builder()
             .chainId('cfb5d93e747d20433e3b14603f90a5eb152d0399e7278f9671ecf9763f8780e8')
             .extId('extId', 'utf8')
@@ -56,7 +57,7 @@ describe('Test Entry', function() {
         assert.deepEqual(entry, copy);
     });
 
-    it('should copy empty Entry', function() {
+    it('should copy empty Entry', function () {
         const entry = Entry.builder().build();
 
         const copy = Entry.builder(entry).build();
@@ -67,7 +68,7 @@ describe('Test Entry', function() {
         assert.deepEqual(copy.content, Buffer.from(''));
     });
 
-    it('should get sizes without ext ids', function() {
+    it('should get sizes without ext ids', function () {
         const entry = Entry.builder()
             .content('abcdef', 'utf8')
             .build();
@@ -77,7 +78,7 @@ describe('Test Entry', function() {
         assert.equal(entry.rawDataSize(), 6);
     });
 
-    it('should get remaining free bytes', function() {
+    it('should get remaining free bytes', function () {
         const e1 = Entry.builder()
             .content(Buffer.alloc(24))
             .build();
@@ -97,7 +98,7 @@ describe('Test Entry', function() {
         assert.equal(e4.remainingFreeBytes(), 1024);
     });
 
-    it('should get remaining maximum bytes', function() {
+    it('should get remaining maximum bytes', function () {
         const e1 = Entry.builder()
             .content(Buffer.alloc(240))
             .build();
@@ -113,7 +114,7 @@ describe('Test Entry', function() {
         assert.throws(() => e3.remainingMaxBytes(), Error);
     });
 
-    it('should get hex values', function() {
+    it('should get hex values', function () {
         const e = Entry.builder()
             .chainId('45f7ebb3be5217d0e2f1d14ab73121a66cdaad12a50b9863a45ee8ee9f3ab032')
             .extId('efef', 'hex')
@@ -127,7 +128,7 @@ describe('Test Entry', function() {
         assert.equal(e.chainIdHex, '45f7ebb3be5217d0e2f1d14ab73121a66cdaad12a50b9863a45ee8ee9f3ab032');
     });
 
-    it('should compute marshal binary', function() {
+    it('should compute marshal binary', function () {
         // Examples from Factom data structures document
         const e = Entry.builder()
             .chainId('954d5a49fd70d9b8bcdb35d252267829957f7ef7fa6c74f88419bdc5e82209f4')
@@ -145,7 +146,7 @@ describe('Test Entry', function() {
         assert.equal(e2.marshalBinary().toString('hex'), '00954d5a49fd70d9b8bcdb35d252267829957f7ef7fa6c74f88419bdc5e82209f400005061796c6f616448657265');
     });
 
-    it('should compute hash', function() {
+    it('should compute hash', function () {
         const e = Entry.builder()
             .chainId('954d5a49fd70d9b8bcdb35d252267829957f7ef7fa6c74f88419bdc5e82209f4')
             .content('PayloadHere', 'utf8')
@@ -156,7 +157,7 @@ describe('Test Entry', function() {
         assert.equal(hash.toString('hex'), '72177d733dcd0492066b79c5f3e417aef7f22909674f7dc351ca13b04742bb91');
     });
 
-    it('should compute EC cost', function() {
+    it('should compute EC cost', function () {
         const e1 = Entry.builder()
             .chainId('954d5a49fd70d9b8bcdb35d252267829957f7ef7fa6c74f88419bdc5e82209f4')
             .content('ab')
@@ -184,7 +185,7 @@ describe('Test Entry', function() {
         assert.equal(e3.ecCost(), e1.ecCost() + 1);
     });
 
-    it('should compose Entry commit', function() {
+    it('should compose Entry commit', function () {
         const e1 = Entry.builder()
             .chainId('954d5a49fd70d9b8bcdb35d252267829957f7ef7fa6c74f88419bdc5e82209f4')
             .content('ab')
@@ -197,7 +198,39 @@ describe('Test Entry', function() {
             '000162a2e0a0c8fd461748e49aa77a6380c04059bd7e3040c9dbceca1828b37ddb737dd928909f015d54e4b02234a10b542573645f7ba55650f25eb931985cddcf451df77594b5b678790d79fc44d3c81fb701b2c3278b50b98eec215cd28be19bba5e1d96f7dc262ec8a87ae4cf9ea20eb43ef196052e761afb02a8f1a78df097c27c56a4a1d00f');
     });
 
-    it('should compose Entry reveal', function() {
+    it('should compose Entry commit with provided signature', function () {
+        const e1 = Entry.builder()
+            .chainId('954d5a49fd70d9b8bcdb35d252267829957f7ef7fa6c74f88419bdc5e82209f4')
+            .content('ab')
+            .timestamp(1523151053000)
+            .build();
+
+        const commit = composeEntryCommit(e1, 'EC2UFobcsWom2NvyNDN67Q8eTdpCQvwYe327ZeGTLXbYaZ56e9QR', '78790d79fc44d3c81fb701b2c3278b50b98eec215cd28be19bba5e1d96f7dc262ec8a87ae4cf9ea20eb43ef196052e761afb02a8f1a78df097c27c56a4a1d00f');
+        assert.instanceOf(commit, Buffer);
+        assert.equal(commit.toString('hex'),
+            '000162a2e0a0c8fd461748e49aa77a6380c04059bd7e3040c9dbceca1828b37ddb737dd928909f015d54e4b02234a10b542573645f7ba55650f25eb931985cddcf451df77594b5b678790d79fc44d3c81fb701b2c3278b50b98eec215cd28be19bba5e1d96f7dc262ec8a87ae4cf9ea20eb43ef196052e761afb02a8f1a78df097c27c56a4a1d00f');
+    });
+
+    it('should reject invalid signature manually provided for entry commit (timestamp not fixed)', function () {
+        const e1 = Entry.builder()
+            .chainId('954d5a49fd70d9b8bcdb35d252267829957f7ef7fa6c74f88419bdc5e82209f4')
+            .content('ab')
+            .build();
+
+        const buffer = composeEntryLedger(e1);
+        const secretKey = sign.keyPair.fromSeed(Buffer.from('19c71085f906291070d07cb64cf160692c033d5e6ac05342f2219c03b4a5c8c6', 'hex')).secretKey;
+        const signature = sign.detached(buffer, secretKey);
+
+        try {
+            composeEntryCommit(e1, 'EC2UFobcsWom2NvyNDN67Q8eTdpCQvwYe327ZeGTLXbYaZ56e9QR', signature);
+        } catch (e) {
+            assert.instanceOf(e, Error);
+            return;
+        }
+        throw new Error('Should have rejected invalid signature');
+    });
+
+    it('should compose Entry reveal', function () {
         const e = Entry.builder()
             .chainId('954d5a49fd70d9b8bcdb35d252267829957f7ef7fa6c74f88419bdc5e82209f4')
             .extId('test', 'utf8')
@@ -209,7 +242,7 @@ describe('Test Entry', function() {
         assert.equal(reveal.toString('hex'), '00954d5a49fd70d9b8bcdb35d252267829957f7ef7fa6c74f88419bdc5e82209f400060004746573745061796c6f616448657265');
     });
 
-    it('should compose Entry commit and reveal', function() {
+    it('should compose Entry commit and reveal', function () {
         const e = Entry.builder()
             .chainId('954d5a49fd70d9b8bcdb35d252267829957f7ef7fa6c74f88419bdc5e82209f4')
             .extId('test', 'utf8')
@@ -224,7 +257,22 @@ describe('Test Entry', function() {
         assert.equal(composed.reveal.toString('hex'), '00954d5a49fd70d9b8bcdb35d252267829957f7ef7fa6c74f88419bdc5e82209f400060004746573745061796c6f616448657265');
     });
 
-    it('should compute entry txId', function() {
+    it('should compose Entry commit and reveal with manually provided signature', function () {
+        const e = Entry.builder()
+            .chainId('954d5a49fd70d9b8bcdb35d252267829957f7ef7fa6c74f88419bdc5e82209f4')
+            .extId('test', 'utf8')
+            .content('PayloadHere', 'utf8')
+            .timestamp(1523151053000)
+            .build();
+
+        const composed = composeEntry(e, 'EC2UFobcsWom2NvyNDN67Q8eTdpCQvwYe327ZeGTLXbYaZ56e9QR', '2a7642156b5cf73ecd0b54b803922c5365d58c372d33353995a1ef5b2ef889bdcd19a3101456379cc339642de1d623f6cade5af10addbcdd589d30e8bfe78702');
+        assert.instanceOf(composed.commit, Buffer);
+        assert.instanceOf(composed.reveal, Buffer);
+        assert.equal(composed.commit.toString('hex'), '000162a2e0a0c8be705a58aea4230e99881f625e74cd085b6ef455b94ff144249b9a2f425e8f96015d54e4b02234a10b542573645f7ba55650f25eb931985cddcf451df77594b5b62a7642156b5cf73ecd0b54b803922c5365d58c372d33353995a1ef5b2ef889bdcd19a3101456379cc339642de1d623f6cade5af10addbcdd589d30e8bfe78702');
+        assert.equal(composed.reveal.toString('hex'), '00954d5a49fd70d9b8bcdb35d252267829957f7ef7fa6c74f88419bdc5e82209f400060004746573745061796c6f616448657265');
+    });
+
+    it('should compute entry txId', function () {
         const entry = Entry.builder()
             .chainId('3b6432afd44edb3086571663a377ead1d08123e4210e5baf0c8f522369079791')
             .extId('extId', 'utf8')